// lib/dtos/booking-dto.ts
export class BookingDTO {
    ID: string;
    FirstName: string;
    Surname: string;
    ArrivalDate: string;
    DepartureDate: string;
    RoomTypeID: string;
    NumberOfRooms: number;
    Cellphone: string;
    Email: string;
    BookingNumber: string;
    IsConfirmed: boolean;
    IsPaid: boolean;
}